refactor(ui): extract chat state contract
