+++ ui/src/pages/chat/steer-lifecycle.ts
@@ -11,7 +11,6 @@ import {
   getChatAttachmentDataUrl,
   releaseChatAttachmentPayloads,
 } from "./attachment-payload-store.ts";
-import type { ChatState } from "./chat-history.ts";
 import {
   clearPendingQueueItemsForRun,
   clearTransientQueuedMessageProjection,
@@ -28,6 +27,7 @@ import {
   type ChatSendAck,
   type TerminalFailureChatSendAck,
 } from "./chat-send-ack.ts";
+import type { ChatState } from "./chat-state-contract.ts";
 import { readChatSessionProjectionScope, reduceChatSessionProjection } from "./history-merge.ts";
 import { hasAbortableSessionRun } from "./run-lifecycle.ts";
 import { scheduleChatScroll, type ChatScrollHost } from "./scroll.ts";
